Move the Sierra numbers into an "identifiers" namespace
