+++ common/source_model/src/main/scala/weco/catalogue/source_model/sierra/AbstractSierraRecord.scala
@@ -1,5 +1,7 @@
 package weco.catalogue.source_model.sierra
 
+import weco.catalogue.source_model.sierra.identifiers.TypedSierraRecordNumber
+
 import java.time.Instant
 
 trait AbstractSierraRecord[Id <: TypedSierraRecordNumber] {
