+++ common/source_model/src/main/scala/weco/catalogue/source_model/sierra/SierraBibRecord.scala
@@ -1,5 +1,7 @@
 package weco.catalogue.source_model.sierra
 
+import weco.catalogue.source_model.sierra.identifiers.SierraBibNumber
+
 import java.time.Instant
 
 case class SierraBibRecord(
