+++ common/source_model/src/main/scala/weco/catalogue/source_model/sierra/SierraItemRecord.scala
@@ -1,8 +1,11 @@
 package weco.catalogue.source_model.sierra
 
 import java.time.Instant
-
 import uk.ac.wellcome.json.JsonUtil._
+import weco.catalogue.source_model.sierra.identifiers.{
+  SierraBibNumber,
+  SierraItemNumber
+}
 
 import scala.util.{Failure, Success}
