+++ common/source_model/src/main/scala/weco/catalogue/source_model/sierra/identifiers/SierraRecordTypes.scala
@@ -1,0 +1,5 @@
+package weco.catalogue.source_model.sierra.identifiers
+
+object SierraRecordTypes extends Enumeration {
+  val bibs, items, holdings, orders, patrons = Value
+}
