+++ common/source_model/src/test/scala/weco/catalogue/source_model/sierra/SierraOrderRecordTest.scala
@@ -2,6 +2,10 @@ package weco.catalogue.source_model.sierra
 
 import org.scalatest.funspec.AnyFunSpec
 import org.scalatest.matchers.should.Matchers
+import weco.catalogue.source_model.sierra.identifiers.{
+  SierraBibNumber,
+  SierraOrderNumber
+}
 
 import java.time.Instant
