+++ common/source_model/src/test/scala/weco/catalogue/source_model/sierra/SierraTransformableTest.scala
@@ -3,6 +3,7 @@ package weco.catalogue.source_model.sierra
 import org.scalatest.funspec.AnyFunSpec
 import org.scalatest.matchers.should.Matchers
 import weco.catalogue.source_model.generators.SierraGenerators
+import weco.catalogue.source_model.sierra.identifiers.SierraItemNumber
 
 class SierraTransformableTest
     extends AnyFunSpec
